+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNext
 
 - Use OpenLogin ID Token as bearer authorization in API requests
+- Update CERE Blockchain RPC URLs
 
 ### v1.19.0